feat(backNewFeature): move component
